+++ doc/version_history.rst
@@ -14,6 +14,12 @@ v4.5.0
     * Add ``ThermalCommandState``.
     * ``System``: rename two thermal system items, leaving deprecated aliases.
 
+* Add ``AlignComponent`` enum class to `enums.LaserTracker`.
+* `enums.HVAC`:
+
+    * Remove ``DynaleneSafetyState``.
+    * Add ``DynaleneTankLevel``.
+
 v4.4.0
 ------